+++ shell/tables.go
@@ -14,7 +14,7 @@ import (
 
 // tablesCommand print all tables name in DB.
 func (c CommandList) tablesCommand(ctx context.Context, s *Shell, _ []string) error {
-	tables, err := s.usecases.sqlite3.TablesName(ctx)
+	tables, err := s.usecases.metadata.TablesName(ctx)
 	if err != nil {
 		return err
 	}